+++ test/kilt/utils_spec.rb
@@ -11,7 +11,7 @@ describe Kilt::Utils do
       database.expects(:setup!)
 
       Kilt::Utils.setup_db
-        
+
     end
   end
 
@@ -36,9 +36,9 @@ describe Kilt::Utils do
           Kilt.config.stubs(:[]).with(environment.to_sym).returns Object.new
           Kilt.config[environment.to_sym].stubs(:[]).with(:db).returns Object.new
 
-          Kilt::RethinkDbDatabase.stubs(:new)
-                                 .with(Kilt.config[environment.to_sym][:db])
-                                 .returns database
+          Kilt::DB::RethinkDb.stubs(:new)
+                             .with(Kilt.config[environment.to_sym][:db])
+                             .returns database
 
           Kilt::Utils.database.must_be_same_as database
 
@@ -52,9 +52,9 @@ describe Kilt::Utils do
           Kilt.config.stubs(:[]).with(environment.to_sym).returns Object.new
           Kilt.config[environment.to_sym].stubs(:[]).with(:db).returns Object.new
 
-          Kilt::RethinkDbDatabase.expects(:new)
-                                 .with(Kilt.config[environment.to_sym][:db])
-                                 .returns database
+          Kilt::DB::RethinkDb.expects(:new)
+                             .with(Kilt.config[environment.to_sym][:db])
+                             .returns database
 
           Kilt::Utils.database.must_be_same_as Kilt::Utils.database
 
@@ -70,9 +70,9 @@ describe Kilt::Utils do
             Kilt.config.stubs(:[]).returns nil
             Kilt.config.stubs(:[]).with(:db).returns Object.new
 
-            Kilt::RethinkDbDatabase.stubs(:new)
-                                   .with(Kilt.config[:db])
-                                   .returns database
+            Kilt::DB::RethinkDb.stubs(:new)
+                               .with(Kilt.config[:db])
+                               .returns database
 
             Kilt::Utils.database.must_be_same_as database
 
@@ -103,9 +103,9 @@ describe Kilt::Utils do
           Kilt.config.stubs(:[]).with(:development).returns Object.new
           Kilt.config[:development].stubs(:[]).with(:db).returns Object.new
 
-          Kilt::RethinkDbDatabase.stubs(:new)
-                                 .with(Kilt.config[:development][:db])
-                                 .returns database
+          Kilt::DB::RethinkDb.stubs(:new)
+                             .with(Kilt.config[:development][:db])
+                             .returns database
 
           Kilt::Utils.database.must_be_same_as database
